frontend(fix): put sign-in/up links in navbar
